[refactor] Unify tun_mtu as u16 throughout the codebase
tun_mtu originates from config.local.tun.mtu (u16) but was immediately
widened to usize and later narrowed back via try_from().expect().

Now tun_mtu stays u16 in DialContext, PeerEntry::try_connect,
and Orchestrator. Conversion to usize happens only at the boundary
where max_udp_payload is computed (tun_mtu + CONNECT_IP_OVERHEAD).

+++ src/events.rs
@@ -34,7 +34,7 @@ pub(crate) struct DialContext {
     /// TUN interface name (for route-probe exclusion).
     pub tun_if: String,
     /// TUN MTU in bytes.
-    pub tun_mtu: usize,
+    pub tun_mtu: u16,
     /// Tuning parameters (timeouts, buffers, congestion control).
     pub tuning: Tuning,
     /// Runtime handle for UDP I/O actors.
